src/site/MusiFetch/requirements.txt
Cython
ffmpeg
matplotlib
numpy
asyncpg
asyncio
scipy
pydub
youtube_dl
Django>=3.0,<4.0
psycopg2
djangorestframework